+++ ColdSoupGenerator/ColdSoupGenerator/AXFDataSegment.h
@@ -9,8 +9,8 @@ public:
 	AXFDataSegment();
 	~AXFDataSegment();
 	void Load(std::ifstream &,size_t,size_t);
-	std::string *GetStr(int index);
-	double GetDouble(int index);
+	std::string *GetStr(size_t index);
+	double GetDouble(size_t index);
 private:
 	std::vector<char> *data;
 };
